requirements_common.txt
# 本地图像/视频推理所需的共享运行时依赖项。
# 请安装 requirements.txt（默认/CPU 环境）或 requirements_gpu.txt（GPU 环境）其中之一，
# 保证项目环境统一

numpy>=1.23.5,<3.0
opencv-python>=4.8.0,<5.0
Pillow>=9.5.0,<12.0
PyYAML>=6.0,<7.0
scipy>=1.10.0,<2.0
matplotlib>=3.7.0,<4.0
pandas>=2.0.0,<3.0
seaborn>=0.12.0,<1.0
tqdm>=4.65.0,<5.0
requests>=2.31.0,<3.0
psutil>=5.9.0,<8.0

# Detection / tracking / depth
ultralytics>=8.2.0,<9.0
transformers>=4.39.0,<5.0
huggingface_hub>=0.22.0,<1.0
safetensors>=0.4.0,<1.0
timm>=0.9.0,<2.0
filterpy>=1.4.5,<2.0
easydict>=1.10,<2.0
lapx>=0.5.5,<1.0
scikit-image>=0.21.0,<1.0
thop>=0.1.1

# Evaluation helpers used by DeepSORT eval scripts
motmetrics>=1.4.0,<2.0

# Optional training / dataset extras can be installed on demand:
# tensorboard>=2.13.0,<3.0
# pycocotools>=2.0

# UI libraries
PyQt5>=5.15.0,<6.0
streamlit>=1.30.0,<2.0
